Fix m.Query becoming unreachable in table.go. Also added a spinner
Because of early return on len(currentRows) > 5, we now Query here on table.go:147

Also added a spinner to visually show the program is updating.

Note that the outside cmd\update.go is the one advancing the frame and resending Query every 100ms. It receives the Query{} from table\update
.go, then sends it back.